requirements.txt
daml_dit_if==0.3.5
slackclient==2.9.2
